nueva version

+++ Pry_FacturacionWeb/database/migrations/2025_07_08_013233_add_is_active_to_users_table.php
@@ -12,16 +12,20 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->boolean('is_active')
-                ->default(true)
-                ->after('password');
+            if (!Schema::hasColumn('users', 'is_active')) {
+                $table->boolean('is_active')
+                    ->default(true)
+                    ->after('password');
+            }
         });
     }
 
     public function down(): void
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->dropColumn('is_active');
+            if (Schema::hasColumn('users', 'is_active')) {
+                $table->dropColumn('is_active');
+            }
         });
     }
 };
